+++ src/Serwisant/SerwisantApi/Types/SchemaMobile/SecretToken.php
@@ -7,6 +7,16 @@ use Serwisant\SerwisantApi\Types;
 
 class SecretToken extends Types\Type
 {
+  /**
+   * @var string
+  */
+  public $subjectType;
+
+  /**
+   * @var string
+  */
+  public $token;
+
   protected function schemaNamespace()
   {
     return 'SchemaMobile';